Refactor of mail parser, split date validation
Made the code less nested and more Pythonic.  There are
improvements to do, such as, decoding headers (subject,
in-reply-to, name) and recoding it into utf-8, more validations
(use the header 'receipt' to get the date, etc.

The date parsing was split in a separate method, which also
detects old messages (<Y2K) that might be broken using only
two digits to represent a year.

Signed-off-by: Germán Poo-Caamaño <gpoo@gnome.org>