database refactored: use placeholders instead of concatenations
It is not recommendable to concatenate sql statements with
variables. It does not allow a proper escaping, it is prone to SQL
Injection attacks, among other issues.

Still there are other issues (improper mix of placeholders with
and without quotations), but this patch makes the output and
processing exactly equal with respect to master.

Signed-off-by: Germán Poo-Caamaño <gpoo@gnome.org>