Fix recording of people fields 'Cc' and 'To'
This fix is required, but introduces a regression. The 'people'
and 'messages_people' tables used to store their data with
quotations (which is wrong).  The 'messages_people' table used
to store values either with or without quotations (which is worse).
Therefore, there were broken reports.

All placeholders are used without quotations (which are not needed)
and the code is consistent.

Every email addresses are inserted only once per message-id into
'messages_people', but it could be more than once if the email
address has been added to 'From', 'To' and/or 'To'.

People using a database popupalted with a previous version might
want to fix their data before using mlstats with this fix.

Fix https://bugzilla.libresoft.es/show_bug.cgi?id=309

Signed-off-by: Germán Poo-Caamaño <gpoo@gnome.org>