Fix archive retrieval when content delivered is not the expected
There are some archive whose filenames end in .gz, but the
content is delivered as text/plain.  For instance, FreeBSD archives
such as: http://lists.freebsd.org/pipermail/freebsd-current/

Currently mlstats relies on the extension for the transfer.  mlstats
should check the content.

This is a temporary fix (a dirty hack).  Whenever the downloaded
filename is '.gz', it checks whether the content was delivered as
text/plain.  If so, takes the content and store compressed so the
other methods in mlstats do not complain.

Signed-off-by: Germán Poo-Caamaño <gpoo@gnome.org>