Fix local analysis and report
When the Postgresql support was added, only __analyze_remote
was updated, adding an extra argument to __analyze_list_of_files,
which is also called by __analyze_non_remote also call.

This patch add the new argument missing as well as leave the
related part of code similar to __analyze_remote.

Signed-off-by: Germán Poo-Caamaño <gpoo@gnome.org>