Use a custom mailbox class with a file object. Fix #24
Introduced a code deprecated in Python 2.7 (and removed from
3.x) that handles the mailbox as a file object (file pointer). This
should allow us to parse streams, and avoid uncompressing and
storing files.

It was removed from Python because it does not allow to modify
a mbox. However, we don't modify a mbox. So, this fits with the
purpose of mlstats.

As a consequence, this patch also solves the issue of not
processing the 'received' header (Fix #24).