Make url and list name consistent. Fix #12
Make the url stored always be the same regardless if the user
entered with or without trailing slash.  Also takes care or giving
the list alias (or name) a meaningful text (non-empty) as much
as possible.

Add testcases for the class MailingList, which handles the
location, mailing list name, etc.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>