Archive only compressed files. Fix #2
The tool stored mbox files twice: compressed and uncompressed.
With this commit, downloaded contents are only stored compressed
(even when they are uncompressed on the origin).

MBoxArchive instance provides a new property container() which
returns a container to the compressed archive. This container
is a file descriptor that will decompress the content on the fly.
This means that MailArchiveAnalyzer does not recive a filepath to the
mbox anymore. It gets a a MBoxArchive object to parse the content from
it.

Unit tests were also updated.