Decouple the archive retrieval process from the application
Move classes related to Mail Archiving and the archive retrieval
process out of the Application class.

Create separate classes to handle and retrieve from different sources:
local files, gmane, and mailman.

In the class for gmane, uses a generic checker for the latest offset
retrieved based on the file names where the archives are stored locally.