tests: Add basic unit test for multiple databases
Add basic testins for Sqlite, PostgreSQL and MySQL.

For running the tests locally, it is required to define the environment
variable DB with a corresponding database engine (sqlite, postgres or
mysql).

Optionally, DBUSER and DBPASSWORD may be used to set an user and
password to get access to the db. Otherwise, it will be tried with
the current user and without password.